Show root warning on SplashActivity
